+++ app/src/main/AndroidManifest.xml
@@ -15,17 +15,17 @@
         android:theme="@style/Theme.ShoppingCart"
         tools:targetApi="31">
         <activity
-            android:name=".CartActivity"
+            android:name=".cart.CartActivity"
             android:exported="false"
             android:label="@string/title_activity_cart"
             android:theme="@style/Theme.ShoppingCart" />
         <activity
-            android:name=".ProductDetailActivity"
+            android:name=".productdetail.ProductDetailActivity"
             android:exported="false"
             android:label="@string/title_activity_product_detail"
             android:theme="@style/Theme.ShoppingCart" />
         <activity
-            android:name=".ProductListActivity"
+            android:name=".productlist.ProductListActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.ShoppingCart">
