+++ app/src/main/java/nextstep/shoppingcart/productdetail/component/ProductDetailContent.kt
@@ -1,4 +1,4 @@
-package nextstep.shoppingcart.ui.component
+package nextstep.shoppingcart.productdetail.component
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.aspectRatio
