+++ app/src/main/java/nextstep/shoppingcart/productlist/ProductListActivity.kt
@@ -1,12 +1,13 @@
-package nextstep.shoppingcart
+package nextstep.shoppingcart.productlist
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import nextstep.shoppingcart.model.productList.ProductListUiState
+import nextstep.shoppingcart.cart.CartActivity
 import nextstep.shoppingcart.model.productTestDataList
-import nextstep.shoppingcart.ui.ProductListScreen
+import nextstep.shoppingcart.productdetail.ProductDetailActivity
+import nextstep.shoppingcart.productlist.model.ProductListUiState
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
 
 class ProductListActivity : ComponentActivity() {
