+++ app/src/main/java/nextstep/shoppingcart/productlist/ProductListScreen.kt
@@ -1,4 +1,4 @@
-package nextstep.shoppingcart.ui
+package nextstep.shoppingcart.productlist
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
@@ -10,9 +10,9 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import nextstep.shoppingcart.model.productList.ProductListUiState
 import nextstep.shoppingcart.model.productTestDataList
-import nextstep.shoppingcart.ui.component.ProductListContents
+import nextstep.shoppingcart.productlist.component.ProductListContents
+import nextstep.shoppingcart.productlist.model.ProductListUiState
 import nextstep.shoppingcart.ui.component.ProductListTopBar
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
